requirements.txt
flask-sqlalchemy==2.3.1
flask==1.0.2
Werkzeug==0.14
pyJWT==1.4.0